Fixed issues with api.js (Listens on the correct port)
