+++ wikidata/Dockerfile
@@ -17,4 +17,4 @@ COPY . .
 EXPOSE 8004
 
 # Define the command to run your app
-CMD ["node", "updateQuestions.js"]
+CMD ["node", "server.js"]
